build(app/root): move next.js app into app folder
Moves Next.js app into it's own folder so the commit script lies in the root of the repository and
the CodeCommit repo can upload the code into it from that directory when deployed. When Next.js
lives in the root, the CDK falls into recursion and fails to complete since the infra folder also
lives within the root.